refactor: MoveRule패키지 관리
