+++ src/main/java/janggi/domain/moveRules/outOfPalaceMoveRules/OnceMoveRule.java
@@ -1,10 +1,11 @@
-package janggi.domain.moveRules;
+package janggi.domain.moveRules.outOfPalaceMoveRules;
 
 import janggi.domain.Direction;
 import janggi.domain.Piece;
 import janggi.domain.PieceType;
 import janggi.domain.Position;
 import janggi.domain.Team;
+import janggi.domain.moveRules.MoveRule;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
