+++ src/main/java/janggi/domain/moveRules/outOfPalaceMoveRules/PoMoveRule.java
@@ -1,9 +1,10 @@
-package janggi.domain.moveRules;
+package janggi.domain.moveRules.outOfPalaceMoveRules;
 
 import janggi.domain.Direction;
 import janggi.domain.Piece;
 import janggi.domain.Position;
 import janggi.domain.Team;
+import janggi.domain.moveRules.MoveRule;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
