+++ src/main/java/janggi/domain/moveRules/palaceMoveRule/ChaPalaceMoveRule.java
@@ -1,10 +1,11 @@
-package janggi.domain.moveRules;
+package janggi.domain.moveRules.palaceMoveRule;
 
 import janggi.domain.Direction;
 import janggi.domain.Palace;
 import janggi.domain.Piece;
 import janggi.domain.Position;
 import janggi.domain.Team;
+import janggi.domain.moveRules.outOfPalaceMoveRules.ChaMoveRule;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
