+++ src/main/java/janggi/domain/moveRules/palaceMoveRule/PoPalaceMoveRule.java
@@ -1,9 +1,10 @@
-package janggi.domain.moveRules;
+package janggi.domain.moveRules.palaceMoveRule;
 
 import janggi.domain.Palace;
 import janggi.domain.Piece;
 import janggi.domain.Position;
 import janggi.domain.Team;
+import janggi.domain.moveRules.outOfPalaceMoveRules.PoMoveRule;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
