+++ src/test/java/janggi/domain/moveRules/outOfPalaceMoveRules/PoMoveRuleTest.java
@@ -1,4 +1,4 @@
-package janggi.domain.moveRules;
+package janggi.domain.moveRules.outOfPalaceMoveRules;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -8,6 +8,7 @@ import janggi.domain.PieceType;
 import janggi.domain.Position;
 import janggi.domain.Team;
 import janggi.domain.board.Board;
+import janggi.domain.moveRules.MoveRule;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
